requirements.txt
# asyncpg
ruff
# pandas
openpyxl
aiogram
APScheduler
python-decouple